==> scripts/04_Build-Log-Reg-Model-for-CDI-Risk.Rmd <==
---
title: "Build Logistic Regression Model for CDI Risk"
author: "Steph Reynolds"
date: "`r format(Sys.time(), '%b %d, %Y  %H:%M %p')`"
output:
  html_document:
    toc: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Background

## Project

MINDSCAPE: Modeling of infectious network dynamics for surveillance, control and prevention enhancement

## Description

This file reads in `df.CDI` and checks the assumptions for creating a logistic regression model to evaluate C diff risk among hospitalized patients.

## Outcome/Dependent Variable:

*Outcome is binary, thus, logistic regression is appropriate for assessing risk.*

-   `HACDIF` --\> whether patient is diagnosed with hospital- acquired C diff
    -   In other words, whether pt obtained C diff \> 48 hours after hospital admission

## Predictors/Independent Variables:

*All predictors are binary.*

-   `ID` --\> unique de-identified encounter ID
-   `readmit` --\> whether patient had a prior hospital admission in the last 30 days
-   `ED_dispo` --\> whether pt was admitted from Emergency Dept (ED) ???
    -   NOTE: Unclear -- verify w/ Leo what this should be to indicate that a patient was admitted to Emergency Dept.
-   `age_gte_65` --\> whether pt is aged 65 years or older at time of admission
-   `PPI` --\> whether pt received proton pump inhibitor (PPI) during hospital stay
-   `GAS` --\> whether pt received gastric acid suppressants (GAS) during hospital stay
-   `anyGAS` --\> whether pt received PPI and/or GAS during hospital stay
-   `LAX` --\> whether pt received laxatives (LAX) during hospital stay
-   `ABX` --\> whether pt received antibiotics during hospital stay
-   `DBM` --\> whether pt is diagnosed with diabetes mellitus (DBM)
    -   NOTE: Excluded as of now since we only have DBM diagnoses for Covid-positive patients.

# Load required packages

```{r Load required packages}
library(tidyverse)
library(corrplot) #plot correlation matrix to assess multicollinearity of IVs 
library(pROC) #calculate AUROC and plot ROC
```

# Import `df_CDI_risk_model.csv` and select only necessary variables

```{r Import data}
df <- read_csv("/Users/sreynolds2/Documents/GitHub/MS-CDI_Risk/MS-CDI_Risk/data/clean/df_CDI_risk_model.csv") %>% 
  select(-c(ID, ED_dispo))
```

# Check Assumptions of Logistic Regression

**Source:** <https://www.lexjansen.com/wuss/2018/130_Final_Paper_PDF.pdf>

## Appropriate Outcome Structure

Binary log reg requires DV to be binary. Whether a pt is diagnosed with C diff is binary.

## Observation Independence

Observations are independent from each other. Each patient is unique. Not matched or repeated measures data.

## Absence of Multicollinearity

There should be little or no multicollinearity among IVs. We can test this assumption with a correlation matrix. Make sure that each combination of variables has a correlation coefficient of *less than 0.80*.

### Correlation matrix

```{r Correlation matrix}
# Create correlation matrix to check whether indep vars are correlated with each other
corrplot(cor(df), method = 'number', type = 'lower') 
```

As per matrix, variables `ABX` and `AMR` are highly correlated (coefficent of 0.92). Thus, we should **NOT** include both of these variables in our model.

Additionally, `anyGAS` and `PPI` are moderately correlated (coefficient of 0.74). We should consider including one of these variables in our model. Another measure we can use to assess multicollinearity is variance inflation factor (VIF).

## Linearity of Independent Variables and Log Odds

Continuous IVs must be linearly related to the logit of the outcome. We can test this by visualizing scatterplot between each predictor and logit values. Since all IVs are binary, can ignore this step.

## Large Sample Size

General guideline is to have at least 10 cases with least frequent outcome for each IV/predictor variable in your model.

Since our model includes **8 predictors** and the probability of HA-CDI = 329/55,037 = **0.006**, then we would like a sample size of about (10\*7/.006) = **11,667**. If we include `ED_dispo` and `DBM`, we would have 10 predictors, and thus, would aim for a sample size of about 16,667. Thus, our model easily meets this requirement.

# Full model

## Build logistic regression model

```{r Build full log reg model}
# Full model 
m1 <- glm(data = df, formula = HACDIF ~ readmit + age_gte_65 + PPI + GAS + anyGAS + LAX + ABX, family = binomial)
summary(m1)

# Full model without `anyGAS` - using complimentary log log (cloglog) method
#m1c <- glm(data = df, formula = HACDIF ~ readmit + age_gte_65 + PPI + GAS + LAX + ABX, family = binomial(link = "cloglog"))
#summary(m1c)
```

All predictors except for `anyGAS` and `LAX` were found to be significant (*p\<0.05*) in the full model.

## Calculate AUROC to evaluate performance

```{r Calculate AUROC for m1}
logit_prob <- predict(m1, newdata = df, type = 'response') 
logit_roc <- pROC::roc(df$HACDIF ~ logit_prob, plot = T, print.auc = T)
```

The **AUROC is 0.82**, indicating strong predictive value. Since our model suggests `anyGAS` and `LAX` to be the least significant predictors, we will create a second model without them.

# Model 2

## Build logistic regression model, excluding `anyGAS` and `LAX`

```{r Build log reg model 2}
m2 <- glm(data = df, formula = HACDIF ~ readmit + age_gte_65 + PPI + GAS + ABX, family = binomial)
summary(m2)
```

All predictors were found to be significant (*p\<0.05*) in model 2.

## Calculate AUROC to evaluate performance

```{r Calculate AUROC for m2}
logit_prob <- predict(m2, newdata = df, type = 'response') 
logit_roc <- pROC::roc(df$HACDIF ~ logit_prob, plot = T, print.auc = T)
```

Even after excluding `anyGAS` and `LAX`, the **AUROC is 0.82**. This shows that the model is very accurate at predicting whether a patient will be diagnosed with hospital-acquired *C. difficile* given the following risk factors: older age, readmission in the last 30 days, and receipt of proton pump inhibitor, gastric acid suppressant, and/or antibiotic.
